Actual working -l
While it was in the help and documents, when finally tested, this functionality didn't actually work, and it was also written as if it took an argument, therefore in it's intended use it would actually absorb other arguments.

This commit actually makes this flag work properly and adds the functionality in the code to make this flag do the thing it was supposed to do.