Add Cache class
This class enables caching of emoji files to speed up repeated runs when
the emoji files are not changed.

The caching is done by placing cache files in a provided cache
directory. The files are further split into directories by their export
format (such as "png-128", "svg", etc).

The individual cache files have a name that is the result of hashing
input data that make such name unique:
  - The contents of the source SVG file;
  - The colours that are applied in the palette being exported, if any.
This allows a cache entry to be invalidated by any changes to the
content of the source SVG file, but also allows targetted invalidation
when manifest colours change.

Moreover, the cache does not depend on details that do not affect the
rendered image; a change in short code, codepoits, description, etc will
not invalidate the cache.
This means that repeated exports that only change the output file naming
structure (such as exporting by codepoint, and exporting by shortcode)
will share the cache.