Fix a major bug with the EXIF step not passing all exports to exiftool
There was a bug that caused a very low number of exports and cached
exports to be reported as being copied to cache after the license (exif)
step ran.

This was the result of an error introduced in 03cd0280.

The `get_formats_for_license_type` returned a generator that was being
consumed in the first format check of the first emoji in the chain,
making the check incomplete for the next formats and emoji, and thus
ignoring other formats and emoji in subsequent checks.
This would make it such that at most only one emoji per format exported
would be passed to exiftool for processing (one of each format, if the
generator order coincided with the format order in the formats tested).

This was fixed by always returning a tuple with the formats from the
`get_formats_for_license_type` function, instead of relying on a
generator.