Skip SVG and SVGO caching when exporting without license
When using orxporter with the `-l` flag (to disable licenses), a bug
existed when the `Cache.save_to_cache` function was called for the SVG
and SVGO formats, as these were (wrongly) being marked as always having
a license (which is the case when `-l` is not specified), thus making
the function look into a property (`licenses`) of the emoji's cache
keys to determine the cache path, but that property was set to `None`
(usually a `dict` when exporting with licenses).

This was resolved by making non-licensed ("exported") SVG files
explicitly non-cacheable, and properly checking for the license flag
(`license_enabled`) when an SVG/SVGO file is being saved to the cache.

Furthermore simply fixing the call of `save_to_cache` after export to
take into account `license_enabled` for SVG and SVGO exports wouldn't be
sufficient.
This is because orxporter only applies licenses after export for EXIF
licenses, not SVG; any later run of orxporter without `-l` could pick up
the non-licensed SVGs by copying them to the output directory (as it's
done for PNGs and other EXIF-enabled formats), but those files would not
go through a step to add license to them, ending up in the output
directory without any license.
For this reason it is recommended to delete all caches of SVG and SVGO
files.